fix(nvim): Move lazy lockfile to data dir
Having the lockfile in the config dir causes permission issues if the
config dir is a symlink to /usr/share/dotfiles/.config/nvim.

It might be worth adding the lockfile to the PKGBUILD, though I'd still
need to find away around the permission errors when updating it. I might
explore that later.