fix(eww): handle destroyworkspacev2 event
Previously, when moving between workspaces very quickly it was possible
to create a "ghost" workspace in the eww bar that wasn't really open.

I think it's still possible to create a "ghost" workspace now (i.e. if
the activateworkspacev2 and destroyworkspacev2 both finish before we can
call hyprctl), but it's pretty unlikely. The correct way to handle it
would probably to properly parse the events coming from the socket, but
this works for now.